+++ beets/autotag/hooks.py
@@ -74,8 +74,11 @@ class AlbumInfo(AttrDict):
             album_id: Optional[str] = None,
             artist: Optional[str] = None,
             artist_id: Optional[str] = None,
+            artists: Optional[List[str]] = None,
+            artists_ids: Optional[List[str]] = None,
             asin: Optional[str] = None,
             albumtype: Optional[str] = None,
+            albumtypes: Optional[List[str]] = None,
             va: bool = False,
             year: Optional[int] = None,
             month: Optional[int] = None,
@@ -83,6 +86,7 @@ class AlbumInfo(AttrDict):
             label: Optional[str] = None,
             mediums: Optional[int] = None,
             artist_sort: Optional[str] = None,
+            artists_sort: Optional[List[str]] = None,
             releasegroup_id: Optional[str] = None,
             release_group_title: Optional[str] = None,
             catalognum: Optional[str] = None,
@@ -96,6 +100,7 @@ class AlbumInfo(AttrDict):
             albumdisambig: Optional[str] = None,
             releasegroupdisambig: Optional[str] = None,
             artist_credit: Optional[str] = None,
+            artists_credit: Optional[List[str]] = None,
             original_year: Optional[int] = None,
             original_month: Optional[int] = None,
             original_day: Optional[int] = None,
@@ -110,9 +115,12 @@ class AlbumInfo(AttrDict):
         self.album_id = album_id
         self.artist = artist
         self.artist_id = artist_id
+        self.artists = artists or []
+        self.artists_ids = artists_ids or []
         self.tracks = tracks
         self.asin = asin
         self.albumtype = albumtype
+        self.albumtypes = albumtypes or []
         self.va = va
         self.year = year
         self.month = month
@@ -120,6 +128,7 @@ class AlbumInfo(AttrDict):
         self.label = label
         self.mediums = mediums
         self.artist_sort = artist_sort
+        self.artists_sort = artists_sort or []
         self.releasegroup_id = releasegroup_id
         self.release_group_title = release_group_title
         self.catalognum = catalognum
@@ -133,6 +142,7 @@ class AlbumInfo(AttrDict):
         self.albumdisambig = albumdisambig
         self.releasegroupdisambig = releasegroupdisambig
         self.artist_credit = artist_credit
+        self.artists_credit = artists_credit or []
         self.original_year = original_year
         self.original_month = original_month
         self.original_day = original_day
@@ -190,14 +200,18 @@ class TrackInfo(AttrDict):
             release_track_id: Optional[str] = None,
             artist: Optional[str] = None,
             artist_id: Optional[str] = None,
+            artists: Optional[List[str]] = None,
+            artists_ids: Optional[List[str]] = None,
             length: Optional[float] = None,
             index: Optional[int] = None,
             medium: Optional[int] = None,
             medium_index: Optional[int] = None,
             medium_total: Optional[int] = None,
             artist_sort: Optional[str] = None,
+            artists_sort: Optional[List[str]] = None,
             disctitle: Optional[str] = None,
             artist_credit: Optional[str] = None,
+            artists_credit: Optional[List[str]] = None,
             data_source: Optional[str] = None,
             data_url: Optional[str] = None,
             media: Optional[str] = None,
@@ -220,6 +234,8 @@ class TrackInfo(AttrDict):
         self.release_track_id = release_track_id
         self.artist = artist
         self.artist_id = artist_id
+        self.artists = artists or []
+        self.artists_ids = artists_ids or []
         self.length = length
         self.index = index
         self.media = media
@@ -227,8 +243,10 @@ class TrackInfo(AttrDict):
         self.medium_index = medium_index
         self.medium_total = medium_total
         self.artist_sort = artist_sort
+        self.artists_sort = artists_sort or []
         self.disctitle = disctitle
         self.artist_credit = artist_credit
+        self.artists_credit = artists_credit or []
         self.data_source = data_source
         self.data_url = data_url
         self.lyricist = lyricist