+++ test/test_plugin_mediafield.py
@@ -33,6 +33,13 @@ field_extension = mediafile.MediaField(
     mediafile.ASFStorageStyle('customtag'),
 )
 
+list_field_extension = mediafile.ListMediaField(
+    mediafile.MP3ListDescStorageStyle('customlisttag'),
+    mediafile.MP4ListStorageStyle('----:com.apple.iTunes:customlisttag'),
+    mediafile.ListStorageStyle('customlisttag'),
+    mediafile.ASFStorageStyle('customlisttag'),
+)
+
 
 class ExtendedFieldTestMixin(_common.TestCase):
 
@@ -58,6 +65,22 @@ class ExtendedFieldTestMixin(_common.TestCase):
         finally:
             delattr(mediafile.MediaFile, 'customtag')
             Item._media_fields.remove('customtag')
+
+    def test_extended_list_field_write(self):
+        plugin = BeetsPlugin()
+        plugin.add_media_field('customlisttag', list_field_extension)
+
+        try:
+            mf = self._mediafile_fixture('empty')
+            mf.customlisttag = ['a', 'b']
+            mf.save()
+
+            mf = mediafile.MediaFile(mf.path)
+            self.assertEqual(mf.customlisttag, ['a', 'b'])
+
+        finally:
+            delattr(mediafile.MediaFile, 'customlisttag')
+            Item._media_fields.remove('customlisttag')
 
     def test_write_extended_tag_from_item(self):
         plugin = BeetsPlugin()